refactor: 재사용성을 위해 spinner 컴포넌트 분리
